Move pretty_printer to own file, maybe speeding up some things
